+++ docs/changelog.md
@@ -9,6 +9,7 @@ Change Log
 Change: kotlinx-metadata 0.9.0. Note that the `KotlinClassMetadata .read` is deprecated in 0.9.0 and replaced with `readStrict` (#1830).
 * Fix: `KSAnnotation.toAnnotationSpec` writes varargs in place instead of making them an array to work around a Kotlin
   issue with `OptIn` annotations (#1831).
+* Fix: `MemberName`s without a package are now correctly imported (#1841)
 
 ## Version 1.16.0